+++ src/components/common/header/Header.tsx
@@ -24,7 +24,7 @@ const Header = () => {
       <div className="flex items-center gap-10 mx-auto w-[1280px] h-full text-text-100">
         <Link href={"/"} className="flex items-center gap-3 w-fit h-full">
           <Image src={Logo} alt="로고" className="w-auto h-full" />
-          <p className="text-title/32px font-DungGeunMo">Good_Game</p>
+          <h1 className="text-title/32px font-DungGeunMo">Good_Game</h1>
         </Link>
         <div
           onClick={onClickIsSelectViewToggleHandler}
